JSON Comparisons
Adds support for JSON dict/list comparisons to
simplify matching between those in results.
Also refactored imports to be more specific on
the `F401` rule and added `none_of` operator as
well as a reverse of the existing `one_of`
operator. To support the list comparisons, a
root list object was also added that mirrors
the root dictionary one.